Implement Scraper component
We begin by implementing a scraper component built using x-ray and cheerio
which is able to handle multiple pages of a specific brand.
We also try to minimize dependencies with DOM selectors.
Using the parser.js helper method, nutrition information is scraped with as low specificity
as possible to help prevent breaking changes in the future.
Unit tests (using Mocha) are included.